+++ frontend/app/lib/models.ts
@@ -1,0 +1,4 @@
+export interface chartData {
+  name: string;
+  value: number;
+}
